backend/requirements.txt
# Tax Appeal Analysis System - Python Dependencies
# 税务诉求分析系统 - Python 依赖包

# Web UI Framework
streamlit>=1.28.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0
openpyxl>=3.10.0

# NLP & Text Embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Clustering & Dimensionality Reduction
scikit-learn>=1.3.0
umap-learn>=0.5.0
hdbscan>=0.8.0

# API Integration
openai>=1.0.0

# Visualization & Analysis
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bar
tqdm>=4.65.0

# NLP Models (for NER tokenization)
jieba>=0.42.1

# Web API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
